+++ afk_slack_agent/os_interaction_utils.py
@@ -3,10 +3,12 @@ See https://betterprogramming.pub/custom-system-notifications-from-python-mac-5f
 See also https://apple.stackexchange.com/questions/135728/using-applescript-to-lock-screen
 """
 
+import os
 import subprocess
 
 
 def sleep():
+    # On some MacOS version the system geos to sleep very quickly, and the program is halted too quickly
     subprocess.Popen("""osascript -e 'tell application "Finder" to sleep'""", shell=True)
 
 
@@ -17,4 +19,13 @@ def lock_screen():
             """ "q" using {control down, command down}'"""
         ),
         shell=True,
+    )
+
+
+def system_message(message):
+    os.system(
+        """osascript -e 'display notification "{}" with title "{}"'""".format(
+            message,
+            "AFK Agent",
+        )
     )